FW-034 Move language switch code to ThemedAppCompatActivity [GF,SL]
